sentinel-ai/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-ai"
version = "1.0.0"
description = "Autonomous Cyber Threat Hunter with Vision Transformers and Multi-Agent LLM Reasoning"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Sentinel AI Team"}
]
keywords = ["cybersecurity", "threat-hunting", "ai", "vision-transformer", "reinforcement-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "langchain>=0.2.16",
    "langchain-openai>=0.1.25",
    "langchain-anthropic>=0.1.23",
    "openai>=1.45.0",
    "anthropic>=0.34.2",
    "torch>=2.3.1",
    "torchvision>=0.18.1",
    "numpy>=1.26.4",
    "pandas>=2.2.2",
    "scikit-learn>=1.5.1",
    "gymnasium>=0.29.1",
    "stable-baselines3>=2.3.2",
    "elasticsearch>=8.14.0",
    "neo4j>=5.23.1",
    "redis>=5.0.8",
    "scapy>=2.5.0",
    "fastapi>=0.112.2",
    "uvicorn[standard]>=0.30.6",
    "httpx>=0.27.2",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.1",
    "matplotlib>=3.9.2",
    "Pillow>=10.4.0",
    "loguru>=0.7.2",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[project.scripts]
sentinel = "main:main"

[project.urls]
Homepage = "https://github.com/your-org/sentinel-ai"
Documentation = "https://github.com/your-org/sentinel-ai/wiki"
Repository = "https://github.com/your-org/sentinel-ai"
Issues = "https://github.com/your-org/sentinel-ai/issues"

[tool.setuptools]
packages = ["agents", "api", "config", "detection", "graph", "ingestion", "mitigation", "simulation"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short --strict-markers"
markers = [
    "asyncio: mark test as async",
    "slow: mark test as slow (skipped in fast mode)",
    "integration: mark test as requiring live services",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
